introduce timestampSpec + various fixes
