+++ cmds/ocm/commands/ocmcmds/components/sign/cmd_test.go
@@ -274,7 +274,7 @@ created rsa key pair ca.priv[ca.cert]
 
 		// create signing vcertificate from CA
 		buf.Reset()
-		Expect(env.CatchOutput(buf).Execute("create", "rsakeypair", "--ca", "CN=mandelsoft", "C=DE", "--ca-key", "ca.priv", "--ca-cert", "ca.cert", "--root-certs", "root.cert", "key.priv")).To(Succeed())
+		Expect(env.CatchOutput(buf).Execute("create", "rsakeypair", "CN=mandelsoft", "C=DE", "--ca-key", "ca.priv", "--ca-cert", "ca.cert", "--root-certs", "root.cert", "key.priv")).To(Succeed())
 		Expect(buf.String()).To(StringEqualTrimmedWithContext(`
 created rsa key pair key.priv[key.cert]
 `))
