+++ docs/reference/ocm_hash.md
@@ -1,4 +1,4 @@
-## ocm hash &mdash; Hash And Normalization Operations
+## ocm hash &mdash; Commands Acting On Hashes
 
 ### Synopsis
 
@@ -21,5 +21,5 @@ ocm hash [<options>] <sub command> ...
 
 ##### Sub Commands
 
-* [ocm hash <b>componentversions</b>](ocm_hash_componentversions.md)	 &mdash; hash component version
+* ocm hash <b>sign</b>	 &mdash; sign hash
