+++ pkg/contexts/ocm/compdesc/meta/v1/signature.go
@@ -166,14 +166,22 @@ func (s *SignatureSpec) ConvertToSigning() *signing.Signature {
 	}
 }
 
+type TimestampSpec struct {
+	// Value contains the PEM encoded TSA value.
+	Value string     `json:"value"`
+	Time  *Timestamp `json:"time,omitempty"`
+
+	// later, we can add other storage kind, like localBlob.
+}
+
 // Signature defines a digest and corresponding signature, identifiable by name.
 // +k8s:deepcopy-gen=true
 // +k8s:openapi-gen=true
 type Signature struct {
-	Name      string        `json:"name"`
-	Digest    DigestSpec    `json:"digest"`
-	Signature SignatureSpec `json:"signature"`
-	Timestamp string        `json:"timestamp,omitempty"`
+	Name      string         `json:"name"`
+	Digest    DigestSpec     `json:"digest"`
+	Signature SignatureSpec  `json:"signature"`
+	Timestamp *TimestampSpec `json:"timestamp,omitempty"`
 }
 
 // Copy provides a copy of the signature data.
