+++ pkg/contexts/ocm/compdesc/versions/v2/jsonscheme/bindata.go
@@ -78,7 +78,7 @@ func (fi bindataFileInfo) Sys() interface{} {
 	return nil
 }
 
-var _ResourcesComponentDescriptorV2SchemaYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x5a\x5f\x6f\x1b\xb9\x11\x7f\xd7\xa7\x18\x9c\x03\x50\x8e\xbd\x96\xa3\x22\x05\xa2\x17\xc3\xcd\xa1\x40\xd0\xde\xf9\x90\xa4\x7d\xa8\xe3\x06\xd4\xee\x48\x62\xca\x25\x55\x92\x52\xac\xcb\xf9\xbb\x1f\x48\x2e\xf7\xff\xae\xfe\xd9\xb9\x0b\x10\x3f\x24\x22\x77\x38\x33\x1c\xfe\x66\x38\x33\xbb\xcf\x58\x32\x01\xb2\x30\x66\xa9\x27\xa3\xd1\x9c\xaa\x04\x05\xaa\x8b\x98\xcb\x55\x32\xd2\xf1\x02\x53\xaa\x47\xb1\x4c\x97\x52\xa0\x30\x51\x82\x3a\x56\x6c\x69\xa4\x8a\xd6\x63\x32\x78\xe6\x29\x4a\x1c\x3e\x69\x29\x22\x3f\x7b\x21\xd5\x7c\x94\x28\x3a\x33\xa3\xf1\xe5\xf8\x32\x7a\x31\xce\x18\x92\x41\x60\xc3\xa4\x98\x00\xb9\x59\xa2\x80\xd7\x41\x06\xfc\x24\x13\xe4\xb0\x1e\x43\x41\x3d\x63\x82\x59\x62\x3d\x19\x00\xa4\x68\xa8\xfd\x1f\xc0\x6c\x96\x38\x01\x22\xa7\x9f\x30\x36\xc4\x4d\x55\x39\xe7\x8a\x43\xa1\xb8\x5b\x9f\x50\x43\xfd\x02\x85\xff\x5f\x31\x85\x89\xe7\x08\x10\x01\xf1\x72\xff\x8d\x4a\x33\x29\x3c\xd5\x52\xc9\x25\x2a\xc3\x50\x07\xba\x0a\x51\x98\xcc\x55\xd2\x46\x31\x31\x27\x03\xa7\xae\x9a\x63\xa7\xbe\x4d\xc6\x94\xcf\xa5\x62\x66\x91\x16\x4c\x97\xd4\x18\x54\x76\x43\xff\xbd\xa5\xd1\xaf\x77\xf6\x9f\xcb\xe8\xd5\xe8\x63\x74\x77\xf6\x8c\x64\x64\xb1\x14\x33\x36\x9f\xc0\x17\x78\x70\x33\x34\x49\x9c\xcd\x28\xff\xa5\x90\x01\x33\xca\x35\x0e\x00\x38\x9d\x22\xef\xd4\xaa\xc5\x28\x82\xa6\x48\x8a\xe1\x9a\xf2\x15\x76\x6d\xc1\xd2\x76\x9a\xc4\x4f\xba\xf5\x13\xf8\xf2\x10\xc6\x75\x43\x96\xf6\xbc\xbe\xbd\x8c\x5e\x95\x76\xaa\xd9\x5c\x30\x31\x6f\x48\x98\x4a\xc9\x91\x8a\x40\x56\x32\xbc\xfd\x7b\xa6\x70\x36\x01\x72\x32\x2a\xc1\x69\xe4\x68\xdc\x31\xe5\x50\xf9\x39\x57\xbe\x45\xf1\x94\xde\xff\x13\xc5\xdc\x2c\x26\x30\x7e\xf9\x72\xd0\x7a\x38\x91\x3f\x9d\xbb\xe7\xc3\xdb\x8b\xbb\xda\xd4\xe9\xf3\x30\xf7\x65\x7c\xfe\x30\x1c\x55\x1e\x7f\x6c\x59\xf2\xd1\xae\x39\xb5\x7b\x1f\x00\xb0\x04\x85\x61\x66\x73\x6d\x8c\x62\xd3\x95\xc1\x7f\xe0\xc6\xab\x9a\x32\x91\xeb\xd5\xa6\x95\x15\x3e\xbc\x8d\x3e\x9e\x05\x45\xc2\xe4\xe9\x95\x67\xad\x90\xd3\x7b\x4c\xde\x61\xba\x46\xe5\x79\x9e\x80\xa1\xff\x43\x01\x33\x25\x53\xd0\xee\x81\x75\x69\xa0\x22\x01\x9a\x7c\x5a\x69\x83\x09\x18\x09\x94\x73\xf9\x19\xa8\x00\xb9\xf4\x78\x03\x8e\x34\x61\x62\x0e\x64\x4d\xce\x21\xa5\x9f\xa4\x8a\xa4\xe0\x9b\x73\xb7\xd4\x8d\x2f\x52\x26\xb2\xd9\x20\x6b\xc1\x34\xa4\x48\x85\x06\xb3\x40\x98\x49\xcb\xd5\x32\xf1\xe6\xd7\x40\x15\x5a\x51\x16\x39\x2c\xa9\xea\xab\x83\xc2\x2f\x2e\xc6\x17\x7f\x29\xff\x8e\x66\x52\x9e\x4d\xa9\xca\xe6\xd6\x65\x82\x75\x1b\xc5\x8b\x8b\x71\xf8\x95\x93\x95\xe8\xf3\x9f\x95\x65\x65\x63\xaf\xef\xae\x86\x97\xbf\xdd\xbe\x88\x5e\xdd\x7d\x48\x9e\x9f\x0e\xaf\x26\x1f\x2e\xca\x13\xa7\x57\xed\x53\xd1\x70\x78\x35\x29\x26\x7f\xfb\x90\xb8\x33\xba\x8e\xfe\x13\xdd\x59\xfc\x87\xdf\x81\xe5\x8e\xc4\xa7\x41\xe2\xd9\xb0\xfc\xe0\xcc\x31\xa9\xcc\x38\xca\xcc\xc7\x9a\x51\xac\x01\xbd\x6d\x11\x6d\x63\xfd\x48\xdb\x70\xd4\xea\x78\x6d\x50\x26\xf0\xe0\xa1\xb8\x94\x9a\x19\xa9\x36\xaf\xa5\x30\x78\x6f\xf6\x09\x53\x96\xaa\x2b\x2c\x39\x0e\x3d\x91\x5a\xc6\xec\x6d\xbb\x6c\xca\xf9\xcd\xac\x90\xd2\xba\xa3\x86\xda\x45\xb4\xac\xeb\x99\xe9\x3a\xa5\x1a\xff\xa5\x38\x29\x62\x5e\x43\x65\xfb\x97\x91\x95\xa7\x3a\x82\xaa\xbf\x06\x4a\x71\xec\x27\xba\x5c\x56\x22\x65\xef\x52\x00\x14\xab\x74\x02\xb7\x64\xa5\xf8\x2f\xd4\x2c\xc8\x39\x10\xbd\xa0\xe3\x97\x7f\x8d\x12\x36\x47\x6d\xc8\xdd\xa0\xc6\x67\x5f\xce\xce\xc6\x73\xa6\x8d\xda\x58\xee\x37\xaf\xdf\xe4\xc3\x3b\x7b\x06\x34\x8e\x51\xeb\x1d\xaf\x77\x6b\x19\x47\x05\x33\xa9\xb2\xa5\xa8\x61\x68\x47\x78\x6f\x50\xd8\x2b\x45\x9f\x6e\x01\xcb\x00\x60\xce\xcc\x62\x35\xbd\xee\x97\xdd\x8b\x36\x37\xb4\x10\x28\x1d\xa8\x9b\x99\x1d\x84\xc6\xba\xd9\xbc\x82\xb9\xf9\x33\x41\x5b\x96\x5b\x94\xf6\x53\xc4\x32\x4d\x99\xe9\xf3\x09\x21\x05\x1e\x63\x97\x23\xf7\xfd\xb3\x14\xe8\x81\xa1\xe5\x4a\xc5\xf8\x63\xee\x70\x7b\xa8\x63\xd3\x91\x7c\x90\x25\x1a\xf9\xd8\x72\xc8\x07\x1e\x42\x87\x67\x35\x1d\x59\x46\x6b\xb0\xcb\x96\xe0\xbd\x51\xf4\x4d\x46\xb0\x25\x5b\x69\xf0\x21\x5d\xd9\x53\x47\x84\x2a\xdd\x99\x64\xf7\xe3\x70\xb9\xa2\x6e\x10\x51\xa5\xe8\xa6\xd8\x39\x33\x98\x56\xe2\x56\xab\x0e\x8e\x57\x58\x54\x76\x76\x37\x16\x9b\x9b\x59\x35\x48\xb6\x32\xf1\xeb\xc8\x76\xc2\xb2\x5f\xef\x40\x6e\x8b\x98\x40\x3c\x00\xf0\x31\xef\xdd\x12\xe3\x3d\xc0\xb6\xa0\x7a\x71\x1d\x52\xf8\x02\x82\x52\xa5\x94\x33\x4d\xad\xa0\xe6\x63\x97\x0d\x77\xc0\xae\xc2\xb0\x7e\x08\xfe\xa0\x02\x40\x5b\x85\xf4\x2e\xf1\x69\x78\x3b\xc5\xc0\x67\xda\xd4\xac\x14\xee\x69\x04\xda\xb3\x43\x3b\x4a\x31\x61\xf4\x7d\xf0\xbc\x9d\x6a\xa0\x3d\x95\xf7\x53\xb9\x9c\x82\xaa\x7a\x83\xbc\x5f\xa0\x27\xf2\xd7\x88\x9c\xb9\xe4\x33\xdf\x36\x94\xca\x9c\x5e\xfb\x1c\x1a\x8d\x3c\xc4\xf2\x61\xce\xcf\x67\x1d\xbd\x15\xdc\xae\x21\xaa\x62\x10\x2f\x6f\x4b\x9c\x28\x70\x5f\xae\xb8\x4a\xfb\xec\x5c\x59\xc1\x4b\x1e\x61\x58\x8a\xda\xd0\x74\xd9\x8d\x33\x81\xb6\x98\xf8\xf1\x10\x7f\xcb\xcd\x79\x80\x35\x1a\x61\xb3\x85\xe6\x51\xe2\xf3\xde\x66\xcf\x6d\x92\xb7\x45\xbc\x71\xf6\xb9\x85\x3b\xaf\xbd\x6d\x96\x6a\xd5\xae\x92\x57\x3e\xc2\xbd\x73\x20\x12\x15\x66\x89\x40\xd9\x1c\x70\xe4\xa5\x54\x87\x9f\xb3\xbf\x56\xf1\xdb\x90\x40\x6d\xcd\x44\xa9\x4d\xb6\x50\xa1\x88\xd1\x95\xc4\x30\x2c\x7a\x66\x5c\xc6\x94\x9f\x66\x09\x4c\x57\x56\x14\xa0\xf3\x0e\x39\xc6\x46\xaa\x43\x91\xf6\x04\x77\x75\xb9\x39\xf2\x36\xec\xf2\x50\xbb\xe4\x9c\x76\x6d\x34\xb5\xe2\x2e\x02\xb2\xee\x6f\xcf\xb5\xb4\x73\xf6\x83\x76\x5f\xb6\x07\x27\x40\x63\xb3\xa2\x9c\x6f\x26\x85\xa4\xc8\x5d\x21\x9f\x47\xa0\x97\x18\x33\xca\x2d\x56\x8d\x62\xb1\x13\xf2\xed\x26\x88\x4f\x96\xfd\xd5\x23\x80\x14\x58\xcf\xfe\x32\x59\x62\xc5\xf9\x0e\xe9\x5b\x2d\x80\x86\x50\x51\xdc\xff\x7b\x16\x94\x81\x81\xde\xb9\x29\x9a\x61\x12\x4e\xdc\x7a\xe7\xf8\x05\x97\xf3\xac\xc7\xb5\xd2\x06\x52\x6a\xe2\x45\xc9\x19\x74\xa3\x2e\x69\xd6\x96\xdc\xe5\x75\xa5\xa9\x72\x1a\xfc\xbd\x5c\xc9\x77\xe5\x03\xf7\x23\x21\xd6\x33\x2b\x6e\x1f\x7f\x08\x3b\xd7\xaf\x0e\x02\xe4\x1c\x08\xde\x1b\x54\x82\xf2\xbc\x84\xff\x76\x8b\x2a\x19\xb3\xbf\x71\xb9\x7b\x55\xe5\x6c\xf0\x77\xc6\x51\x6f\xb4\xc1\x74\xff\xb5\x37\x6d\x02\x9f\x3a\x7a\xc8\x98\xbd\x49\xe9\xfc\xa8\xe6\x87\x1b\x32\xcb\x25\xbf\x37\x1f\xa5\x2b\x52\x6e\xa2\x05\x3c\x55\xc5\x6c\x69\x73\x16\xe6\x3c\x62\x63\x9c\x6e\x82\x5f\x1e\xb7\x1f\x20\x99\x4a\x04\x8a\x06\xd7\xac\xab\x64\xbb\xb6\x1b\xa8\xa6\x15\xb6\x66\x4b\xa9\x60\x33\xd4\xa6\x5e\xac\xd5\x84\x1e\x58\x11\x7a\xcb\xf8\x00\xee\x1d\xc5\x6b\xa0\xc1\xc8\x2d\x12\xeb\x40\x6d\x8a\xf3\x14\x41\x94\xa1\x6a\x8e\x06\x13\x88\xa5\x30\x79\xa2\xd4\xc9\x5e\xb3\x5f\x7b\xf7\x62\x9f\x03\x13\x30\xdd\x18\xd4\x41\xc6\xd4\x1a\xbb\xce\x57\xac\xd2\xa9\x3d\xd0\x01\x40\xa7\xcb\x1e\x01\x97\x19\xe3\x58\xdc\x97\xc7\x22\xa6\x45\xc3\x02\x3d\x41\x54\x97\x5d\xc2\xf3\xb2\x39\xc0\x2c\xa8\x01\xa6\xdd\xde\xad\xf9\x99\x70\xcf\x7e\xb0\x0f\xf5\x0f\x90\x30\xe5\x12\xf3\x4d\xe7\x79\x04\xbb\xdd\x3c\x92\x7f\x3d\x81\xc1\x6e\xea\x7e\xd6\x0f\xce\x2a\x30\x9d\xbf\xc3\x67\x66\x16\x99\x69\xe2\x95\x52\x28\x0c\xb4\xbd\x65\xef\xb3\x52\x08\xad\x6f\xb3\xcc\x68\x1f\x1b\x75\x64\x5c\x9d\x46\xfc\x9e\x23\x6d\xbf\x4b\xdc\x61\x7c\xfd\xc4\xa4\x2b\xb9\x28\x5d\xbb\x5f\xe7\xb2\x1f\x00\x14\x9d\xdf\x23\x1c\x76\x15\x5e\xfd\x1c\x79\xbd\x5b\x65\xf2\xe3\x58\xf5\xbc\xe6\x19\x00\xcc\x51\xa0\x62\xf1\x1f\xf8\x8a\x26\xd3\xc0\xbf\xa5\xc9\x06\xdf\x3d\xfb\x4f\xe0\xd9\xc5\xc1\xf8\xf9\x3f\xd6\xb1\x2b\x40\xfd\x5a\x49\x7c\x7e\x33\xed\xdc\xae\xda\xbb\x3f\xd5\xc4\x69\xe3\x43\x00\x5d\x7a\xb8\x54\x72\xcd\x92\xe2\x44\x23\x20\x95\x26\x43\xb5\xe7\x95\xe7\xf3\xba\xc2\xbf\xb2\xe2\x4f\xd1\xcd\x8d\x15\xba\xc2\xf8\x3d\x6b\xba\xdd\x6d\x40\xe8\x79\x76\x8c\xc5\x47\x04\x33\xa9\x52\x6a\x26\x90\x50\x83\x91\x61\x79\xbf\xba\x69\xc2\xbd\x61\xdb\x28\x7b\xfb\xea\xd9\xc6\x67\x1f\x04\x4e\x42\x7a\xc3\x37\xe7\xf0\x19\x41\x0a\xbe\xc9\x3e\x75\x72\x55\x80\x14\x41\xd9\x70\xa4\x5b\x1c\xf3\xc9\xdc\x2f\x43\xc3\x23\xf5\x3b\x6a\xaf\xd9\xf3\x03\x6e\x42\xf2\x71\x04\x36\x19\xd7\x5b\xfd\x4f\x79\xf6\xe5\x1e\x21\xd9\x11\x2c\x95\xdc\x75\xa7\x45\xb5\x5b\xd1\x85\xa6\x76\x93\xc2\x97\x87\xc1\x60\x50\x8b\x53\xe5\x20\x14\x01\x49\xd1\x7f\xbb\x5a\x0e\x14\x64\x50\x0d\x03\xc5\x37\xb2\x1d\x9f\x3d\x7a\x16\xb5\xf8\xd8\x7f\x40\xa4\xfc\xc2\xb3\x9a\x6b\x94\x0e\xa4\x72\x18\xfd\x2f\x09\x49\xed\xcd\xdf\x11\x3c\xdb\x5f\x96\x91\xdf\x03\x00\x00\xff\xff\x9e\xf8\x98\x8b\xde\x2c\x00\x00")
+var _ResourcesComponentDescriptorV2SchemaYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xec\x5a\x5f\x6f\x1b\xb9\x11\x7f\xd7\xa7\x18\x9c\x03\x50\x8e\xbd\x96\xa3\x22\x05\xa2\x17\xc3\xcd\xa1\x40\xd0\xde\xf9\x90\xa4\x7d\xa8\xe3\x06\xd4\xee\x48\x62\xba\x4b\xaa\x24\xa5\x58\x97\xf3\x77\x3f\x90\x5c\xee\x72\xff\xea\x9f\x9d\xbb\x00\xf1\x43\x22\x72\x87\x33\xc3\x99\xdf\x0c\x87\xb3\xfb\x8c\x25\x13\x20\x0b\xad\x97\x6a\x32\x1a\xcd\xa9\x4c\x90\xa3\xbc\x88\x53\xb1\x4a\x46\x2a\x5e\x60\x46\xd5\x28\x16\xd9\x52\x70\xe4\x3a\x4a\x50\xc5\x92\x2d\xb5\x90\xd1\x7a\x4c\x06\xcf\x1c\x45\xc0\xe1\x93\x12\x3c\x72\xb3\x17\x42\xce\x47\x89\xa4\x33\x3d\x1a\x5f\x8e\x2f\xa3\x17\xe3\x9c\x21\x19\x78\x36\x4c\xf0\x09\x90\x9b\x25\x72\x78\xed\x65\xc0\x4f\x22\xc1\x14\xd6\x63\x28\xa9\x67\x8c\x33\x43\xac\x26\x03\x80\x0c\x35\x35\xff\x03\xe8\xcd\x12\x27\x40\xc4\xf4\x13\xc6\x9a\xd8\xa9\x2a\xe7\x42\x71\x28\x15\xb7\xeb\x13\xaa\xa9\x5b\x20\xf1\xff\x2b\x26\x31\x71\x1c\x01\x22\x20\x4e\xee\xbf\x51\x2a\x26\xb8\xa3\x5a\x4a\xb1\x44\xa9\x19\x2a\x4f\x57\x21\xf2\x93\x85\x4a\x4a\x4b\xc6\xe7\x64\x60\xd5\x95\x73\xec\xd4\xb7\xc9\x98\xa6\x73\x21\x99\x5e\x64\x25\xd3\x25\xd5\x1a\xa5\xd9\xd0\x7f\x6f\x69\xf4\xeb\x9d\xf9\xe7\x32\x7a\x35\xfa\x18\xdd\x9d\x3d\x23\x39\x59\x2c\xf8\x8c\xcd\x27\xf0\x05\x1e\xec\x0c\x4d\x12\x6b\x33\x9a\xfe\x52\xca\x80\x19\x4d\x15\x0e\x00\x52\x3a\xc5\xb4\x53\xab\x16\xa3\x70\x9a\x21\x29\x87\x6b\x9a\xae\xb0\x6b\x0b\x86\xb6\xd3\x24\x6e\xd2\xae\x9f\xc0\x97\x07\x3f\xae\x1b\x32\xd8\xf3\xfa\xf6\x32\x7a\x15\xec\x54\xb1\x39\x67\x7c\xde\x90\x30\x15\x22\x45\xca\x3d\x59\x60\x78\xf3\xf7\x4c\xe2\x6c\x02\xe4\x64\x14\xc0\x69\x64\x69\xac\x9b\x0a\xa8\xfc\x5c\x28\xdf\xa2\x78\x46\xef\xff\x89\x7c\xae\x17\x13\x18\xbf\x7c\x39\x68\x75\x4e\xe4\xbc\x73\xf7\x7c\x78\x7b\x71\x57\x9b\x3a\x7d\xee\xe7\xbe\x8c\xcf\x1f\x86\xa3\xca\xe3\x8f\x2d\x4b\x3e\x9a\x35\xa7\x66\xef\x03\x00\x96\x20\xd7\x4c\x6f\xae\xb5\x96\x6c\xba\xd2\xf8\x0f\xdc\x38\x55\x33\xc6\x0b\xbd\xda\xb4\x32\xc2\x87\xb7\xd1\xc7\x33\xaf\x88\x9f\x3c\xbd\x72\xac\x25\xa6\xf4\x1e\x93\x77\x98\xad\x51\x3a\x9e\x27\xa0\xe9\xff\x90\xc3\x4c\x8a\x0c\x94\x7d\x60\x42\x1a\x28\x4f\x80\x26\x9f\x56\x4a\x63\x02\x5a\x00\x4d\x53\xf1\x19\x28\x07\xb1\x74\x78\x83\x14\x69\xc2\xf8\x1c\xc8\x9a\x9c\x43\x46\x3f\x09\x19\x09\x9e\x6e\xce\xed\x52\x3b\xbe\xc8\x18\xcf\x67\xbd\xac\x05\x53\x90\x21\xe5\x0a\xf4\x02\x61\x26\x0c\x57\xc3\xc4\x99\x5f\x01\x95\x68\x44\x19\xe4\xb0\xa4\xaa\xaf\xf2\x0a\xbf\xb8\x18\x5f\xfc\x25\xfc\x1d\xcd\x84\x38\x9b\x52\x99\xcf\xad\x43\x82\x75\x1b\xc5\x8b\x8b\xb1\xff\x55\x90\x05\xf4\xc5\xcf\xca\xb2\xd0\xd8\xeb\xbb\xab\xe1\xe5\x6f\xb7\x2f\xa2\x57\x77\x1f\x92\xe7\xa7\xc3\xab\xc9\x87\x8b\x70\xe2\xf4\xaa\x7d\x2a\x1a\x0e\xaf\x26\xe5\xe4\x6f\x1f\x12\xeb\xa3\xeb\xe8\x3f\xd1\x9d\xc1\xbf\xff\xed\x59\xee\x48\x7c\xea\x25\x9e\x0d\xc3\x07\x67\x96\x49\x65\xc6\x52\xe6\x31\xd6\xcc\x62\x0d\xe8\x6d\xcb\x68\x1b\x13\x47\xca\xa4\xa3\xd6\xc0\x6b\x83\x32\x81\x07\x07\xc5\xa5\x50\x4c\x0b\xb9\x79\x2d\xb8\xc6\x7b\xbd\x4f\x9a\x32\x54\x5d\x69\xc9\x72\xe8\xc9\xd4\x22\x66\x6f\xdb\x65\xd3\x34\xbd\x99\x95\x52\x5a\x77\xd4\x50\xbb\xcc\x96\x75\x3d\x73\x5d\xa7\x54\xe1\xbf\x64\x4a\xca\x9c\xd7\x50\xd9\xfc\xe5\x64\xe1\x54\x47\x52\x75\xc7\x40\x90\xc7\x7e\xa2\xcb\x65\x25\x53\xf6\x2e\x05\x40\xbe\xca\x26\x70\x4b\x56\x32\xfd\x85\xea\x05\x39\x07\xa2\x16\x74\xfc\xf2\xaf\x51\xc2\xe6\xa8\x34\xb9\x1b\xd4\xf8\xec\xcb\xd9\xda\x78\xce\x94\x96\x1b\xc3\xfd\xe6\xf5\x9b\x62\x78\x67\x7c\x40\xe3\x18\x95\xda\xf1\x78\x37\x96\xb1\x54\x30\x13\x32\x5f\x8a\x0a\x86\x66\x84\xf7\x1a\xb9\x39\x52\xd4\xe9\x16\xb0\x0c\x00\xe6\x4c\x2f\x56\xd3\xeb\x7e\xd9\xbd\x68\xb3\x43\x03\x81\xc0\xa1\x76\x66\x76\x10\x1a\xeb\x66\x73\x0a\x16\xe6\xcf\x05\x6d\x59\x6e\x50\xda\x4f\x11\x8b\x2c\x63\xba\x2f\x26\xb8\xe0\x78\x8c\x5d\x8e\xdc\xf7\xcf\x82\xa3\x03\x86\x12\x2b\x19\xe3\x8f\x45\xc0\xed\xa1\x8e\x29\x47\x8a\x41\x5e\x68\x14\x63\xc3\xa1\x18\x38\x08\x1d\x5e\xd5\x74\x54\x19\xad\xc9\x2e\x5f\x82\xf7\x5a\xd2\x37\x39\xc1\x96\x6a\xa5\xc1\x87\x74\x55\x4f\x1d\x19\x2a\x38\x33\xc9\xee\xee\xb0\xb5\xa2\x6a\x10\x51\x29\xe9\xa6\xdc\x39\xd3\x98\x55\xf2\x56\xab\x0e\x96\x97\x5f\x14\x06\xbb\x1d\xf3\xcd\xcd\xac\x9a\x24\x5b\x99\xb8\x75\x64\x3b\x61\x18\xd7\x3b\x90\x9b\x4b\x8c\x27\x1e\x00\xb8\x9c\xf7\x6e\x89\xf1\x1e\x60\x5b\x50\xb5\xb8\xf6\x25\x7c\x09\x41\x21\x33\x9a\x32\x45\x8d\xa0\xe6\x63\x5b\x0d\x77\xc0\xae\xc2\xb0\xee\x04\xe7\x28\x0f\xd0\x56\x21\xbd\x4b\x5c\x19\xde\x4e\x31\x00\xd0\x2c\x43\xa5\x69\xb6\xac\x1b\xc1\xd9\xa0\x43\xe3\x3e\xa6\xf9\x14\x6b\x86\x52\x85\x00\x4c\x46\xcf\xa8\x9e\x40\x42\x35\x46\x86\xde\x26\x01\x36\xe7\x54\xaf\x24\xee\xe9\x14\xda\x63\x71\x33\xca\x30\x61\xf4\xbd\xcf\x04\x3b\xdd\xc9\xf6\x34\xa6\x9b\x2a\xe4\x94\x54\xd5\x13\xed\xfd\x02\x1d\x91\x3b\xd6\xc4\xcc\x16\xc3\xc5\xb6\x21\xb8\x76\xb5\xf9\xab\x20\x3c\x34\x3b\x3a\xc8\x17\xc3\x82\x9f\xab\x82\x7a\x6f\x94\xbb\xa6\xcc\x8a\x41\x9c\xbc\x2d\x79\xab\x8c\xc3\xf0\x06\x18\xec\xb3\x73\x65\x05\x2f\x24\x00\x9f\x05\xf5\x96\xc5\x15\xf0\xbb\xc3\x10\xcd\x9d\xe7\xc7\x43\xd2\x42\x61\xe5\x03\x8c\xd4\xc8\xee\x2d\x34\x8f\x72\x8c\xec\xed\x8d\xc2\x26\x45\xf7\xc6\x19\x67\x9f\x62\xa1\xf3\x74\xde\x66\xa9\x56\xed\x2a\xe5\xef\x23\x1c\x8f\x07\x02\x54\x62\x5e\xaf\x84\xe6\x80\x23\xcf\xce\x3a\xfc\xac\xfd\x95\x8c\xdf\xfa\x3a\x6f\x6b\xc1\x4c\x4d\x4d\x88\x12\x79\x8c\xf6\xe6\x0e\xc3\xb2\xb5\x97\x8a\x98\xa6\xa7\x79\x9d\xd5\x55\xbc\x79\xe8\xbc\xc3\x14\x63\x2d\xe4\xa1\x48\x7b\x82\x92\x22\xec\xe1\xbc\xf5\xbb\x3c\xd4\x2e\x05\xa7\x5d\xfb\x61\xad\xb8\x8b\x80\xac\xfb\xbb\x88\x2d\x5d\xa7\xfd\xa0\xdd\x57\x94\xc2\x09\xd0\x58\xaf\x68\x9a\x6e\x26\xa5\xa4\xc8\x9e\x2c\x9f\x47\xa0\x96\x18\x33\x9a\x1a\xac\x6a\xc9\x62\x2b\xe4\xdb\xad\x63\x9f\xac\x48\xad\x67\x00\xc1\xb1\x5e\xa4\xe6\xb2\xf8\x2a\x4d\x77\xa8\x32\x6b\x09\xd4\xa7\x8a\xb2\x2c\xd8\xf3\xde\xeb\x19\xa8\x9d\x7b\xb7\x39\x26\xe1\xc4\xae\xb7\x81\x5f\x72\x39\xcf\x5b\x71\x2b\xa5\x21\xa3\x3a\x5e\x04\xc1\xa0\x1a\xd7\xa7\xe6\x15\x38\xb5\xe5\x67\x30\x15\x56\xeb\xdf\x6f\x55\xc5\xae\x5c\xe2\x7e\x24\xc4\x3a\x66\xe5\xe9\xe3\x9c\xb0\xf3\x35\xdb\x42\x80\x9c\x03\xc1\x7b\x8d\x92\xd3\xb4\xe8\x34\x7c\xbb\x77\x3f\x11\xb3\xbf\xa5\x62\xf7\xcb\x9f\xb5\xc1\xdf\x59\x8a\x6a\xa3\x34\x66\xfb\xaf\xbd\x69\x13\xf8\xd4\xd9\x43\xc4\xec\x4d\x46\xe7\x47\xf5\x68\xec\x90\x19\x2e\xc5\xb9\xf9\x28\xcd\x9b\xb0\xd7\xe7\xf1\x54\x15\xb3\xa5\x1b\x5b\x9a\xf3\x88\x8d\xa5\x74\xe3\xe3\xf2\xb8\xfd\x00\xc9\x55\x22\x50\xf6\xe1\x66\x5d\x37\xb9\x6b\xb3\x81\x6a\x59\x61\xae\x72\x19\xe5\x6c\x86\x4a\xd7\xef\x70\x35\xa1\x07\x5e\x14\x9d\x65\x5c\x02\x77\x81\xe2\x34\x50\xa0\xc5\x16\x89\x75\xa0\x36\xc5\x39\x0a\x2f\x4a\x53\x39\x47\x8d\x09\xc4\x82\xeb\xa2\x50\xea\x64\xaf\xd8\xaf\xbd\x7b\x31\xcf\x81\x71\x98\x6e\x34\x2a\x2f\x63\x6a\x8c\x5d\xe7\xcb\x57\xd9\xd4\x38\x74\x00\xd0\x19\xb2\x47\xc0\x65\xc6\x52\x2c\xcf\xcb\x63\x11\xd3\xa2\x61\x89\x1e\x2f\xaa\xcb\x2e\xfe\x79\x68\x0e\xd0\x0b\xaa\x81\x29\xbb\x77\x63\x7e\xc6\xed\xb3\x1f\xcc\x43\xf5\x03\x24\x4c\xda\xc2\x7c\xd3\xe9\x0f\x6f\xb7\x9b\x47\x8a\xaf\x27\x30\xd8\x4d\x3d\xce\xfa\xc1\x59\x05\xa6\x8d\x77\xf8\xcc\xf4\x22\x37\x4d\xbc\x92\x12\xb9\x86\xb6\x8f\x01\xfa\xac\xe4\x53\xeb\xdb\xbc\x32\xda\xc7\x46\x1d\x15\x57\xa7\x11\xbf\xd7\x48\xdb\xcf\x12\xeb\x8c\xaf\x5f\x98\x74\x15\x17\xc1\xb1\xfb\x75\x0e\xfb\x01\x40\xd9\xa0\x3e\x22\x60\x57\xfe\x0d\xd5\x91\xc7\xbb\x51\xa6\x70\xc7\xaa\xe7\x6d\xd4\x00\x60\x8e\x1c\x25\x8b\xff\xc0\x37\x49\xb9\x06\xee\x65\x52\x3e\xf8\x1e\xd9\x7f\x82\xc8\x2e\x1d\xe3\xe6\xff\xd8\xc0\xae\x00\xf5\x6b\x15\xf1\xc5\xc9\xb4\x73\xbb\x6a\xef\xfe\x54\x13\xa7\x8d\xef\x15\x54\xf0\x70\x29\xc5\x9a\x25\xa5\x47\x23\x20\x95\x26\x43\xb5\xe7\x55\xd4\xf3\xaa\xc2\xbf\xb2\xe2\x4f\xd1\xcd\x8d\x25\xda\x8b\xf1\xfb\x96\xf7\x4f\xb7\x1e\xa1\xe7\xb9\x1b\xcb\x6f\x1d\x9a\xef\xa2\xdc\x7c\xd3\x84\x7b\xc3\xb6\x71\xed\xed\xbb\xcf\x36\xbe\x4e\x21\x70\xe2\xcb\x9b\x74\x73\x0e\x9f\x11\x04\x4f\x37\xf9\x17\x59\xf6\x16\x20\xb8\x57\xd6\xbb\x74\x4b\x60\x3e\x59\xf8\xe5\x68\x78\xa4\x7e\x47\xed\x6b\x80\xc2\xc1\x4d\x48\x3e\x8e\xc0\x26\xe3\x7a\xab\xff\x29\x7d\x1f\xf6\x08\xc9\x8e\x60\xa9\xd4\xae\x3b\x2d\xaa\x9d\x8a\x36\x35\xb5\x9b\x14\xbe\x3c\x0c\x06\x83\x5a\x9e\x0a\x93\x50\x04\x24\x43\xf7\x89\x6d\x98\x28\xc8\xa0\x9a\x06\xca\x4f\x79\x3b\xbe\xce\x74\x2c\x6a\xf9\xb1\xdf\x41\x24\x7c\x0f\x5a\xad\x35\x02\x87\x54\x9c\xd1\xff\xee\x90\xd4\xde\xfc\x1d\xc1\xb3\xfd\x65\x19\xf9\x3d\x00\x00\xff\xff\x69\x79\xe9\x8f\x85\x2d\x00\x00")
 
 func ResourcesComponentDescriptorV2SchemaYamlBytes() ([]byte, error) {
 	return bindataRead(
