+++ resources/component-descriptor-v2-schema.yaml
@@ -158,6 +158,15 @@ definitions:
       value:
         type: string
 
+  timestampSpec:
+    type: object
+    properties:
+      value:
+        type: string
+      time:
+        type: string
+        format: date-time
+
   signatureSpec:
     type: 'object'
     required:
@@ -188,7 +197,7 @@ definitions:
       signature:
         $ref: '#/definitions/signatureSpec'
       timestamp:
-        type: string
+        $ref: '#/definitions/timestampSpec'
 
   nestedDigestSpec:
     type: 'object'
